azure-launch-pipeline/pyproject.toml
[project]
name = "seqera-api-demo"
version = "0.1.0"
description = "A mini example of how to use the Seqera API to add a dataset and launch a pipeline"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.19.0",
    "azure-storage-blob>=12.24.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "types-pyyaml>=6.0.12.20241230",
    "types-requests>=2.32.0.20241016",
]
